refactor(connected-apps): rename service to graphQLAgents to match convention
